retro_games_marketplace/requirements.txt
# requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
dnspython==2.4.2
Pillow==10.0.1
python-multipart==0.0.6